various casting fixes

+++ api/oc_knx_sec.c
@@ -525,7 +525,7 @@ oc_core_auth_at_post_handler(oc_request_t *request,
               int64_t *new_array =
                 (int64_t *)malloc(array_size * sizeof(uint64_t));
               if (new_array) {
-                for (int i = 0; i < array_size; i++) {
+                for (size_t i = 0; i < array_size; i++) {
                   new_array[i] = array[i];
                 }
                 g_at_entries[index].ga = new_array;
@@ -1190,7 +1190,7 @@ oc_core_set_at_table(size_t device_index, int index, oc_auth_at_t entry)
       int64_t *new_array = (int64_t *)malloc(array_size * sizeof(uint64_t));
 
       if (new_array) {
-        for (int i = 0; i < array_size; i++) {
+        for (size_t i = 0; i < array_size; i++) {
           new_array[i] = entry.ga[i];
         }
         g_at_entries[index].ga = new_array;
@@ -1202,7 +1202,7 @@ oc_core_set_at_table(size_t device_index, int index, oc_auth_at_t entry)
     oc_at_dump_entry(device_index, index);
   }
   if (index == 0) {
-    // set the oscore stuff
+    // set the OSCORE stuff
   }
 
   return 0;
